fix: link item request modal hooks
